Split oversized modules and cut 0.5.3
Four files crossed the file-size-guard warning during the backlinks work: tree.rs (854 lines), block.rs (843), input.rs (835), lifecycle.rs (669). All four piled multiple responsibilities into one module.

Each split by responsibility into its own directory. Public API unchanged across the lot, verified by paper-verifier (tree CRDT semantics line-by-line vs Kleppmann 2022) and the full CRDT invariant battery (32/32 green). CHANGELOG entry for 0.5.3 lists the unified backlinks pipeline, Insert-mode cross-block nav, the anti-duplication policy, and these splits.
