+++ Cargo.toml
@@ -12,7 +12,7 @@ members = [
 ]
 
 [workspace.package]
-version = "0.5.2"
+version = "0.5.3"
 edition = "2021"
 rust-version = "1.95"
 license = "MIT"
@@ -23,11 +23,11 @@ readme = "README.md"
 
 [workspace.dependencies]
 # Internal crates
-outl-core = { path = "crates/outl-core", version = "0.5.2" }
-outl-md = { path = "crates/outl-md", version = "0.5.2" }
-outl-actions = { path = "crates/outl-actions", version = "0.5.2" }
-outl-exec = { path = "crates/outl-exec", version = "0.5.2" }
-outl-tui = { path = "crates/outl-tui", version = "0.5.2" }
+outl-core = { path = "crates/outl-core", version = "0.5.3" }
+outl-md = { path = "crates/outl-md", version = "0.5.3" }
+outl-actions = { path = "crates/outl-actions", version = "0.5.3" }
+outl-exec = { path = "crates/outl-exec", version = "0.5.3" }
+outl-tui = { path = "crates/outl-tui", version = "0.5.3" }
 
 # IDs and time
 ulid = { version = "1", features = ["serde"] }
