+++ crates/outl-tui/src/input/visual.rs
@@ -1,0 +1,23 @@
+//! Visual mode key handler.
+//!
+//! Visual mode operates on a contiguous range of outline blocks. Keys
+//! that aren't `d`/`x`/`y`/`Tab`/`BackTab` either move the selection
+//! (extending the range) or exit to Normal.
+
+use crate::state::{App, Mode};
+use anyhow::Result;
+use crossterm::event::{KeyCode, KeyEvent};
+
+pub(crate) fn handle_visual_key(app: &mut App, key: KeyEvent) -> Result<()> {
+    match key.code {
+        KeyCode::Esc | KeyCode::Char('V') | KeyCode::Char('v') => app.mode = Mode::Normal,
+        KeyCode::Char('d') | KeyCode::Char('x') => app.delete_visual_range(),
+        KeyCode::Char('y') => app.yank_visual_range(),
+        KeyCode::Tab => app.indent_visual_range(),
+        KeyCode::BackTab => app.outdent_visual_range(),
+        KeyCode::Down | KeyCode::Char('j') => app.move_selection(1),
+        KeyCode::Up | KeyCode::Char('k') => app.move_selection(-1),
+        _ => {}
+    }
+    Ok(())
+}
